+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonIntegratedToolbarPrintCommand.cs
@@ -80,7 +80,7 @@ namespace Krypton.Toolkit
 
             _printButtonSpec = null;
 
-            Text = KryptonLanguageManager.ToolBarStrings.Print;
+            Text = KryptonManager.Strings.IntegratedToolBarStrings.Print;
         }
 
         #endregion