+++ Source/Krypton Components/Krypton.Toolkit/General/Definitions.cs
@@ -1883,4 +1883,39 @@ namespace Krypton.Toolkit
     }
 
     #endregion
+
+    #region Enum AboutToolkitPage
+
+    internal enum AboutToolkitPage
+    {
+        GeneralInformation = 0,
+        Discord = 1,
+        DeveloperInformation = 2,
+        Versions = 3
+    }
+
+    #endregion
+
+    #region Enum AboutBoxFileInformationPage
+
+    public enum AboutBoxFileInformationPage
+    {
+        Application = 0,
+        Assemblies = 1,
+        AssemblyDetails = 2
+    }
+
+    #endregion
+
+    #region Enum AboutBoxPage
+
+    public enum AboutBoxPage
+    {
+        GeneralInformation = 0,
+        Description = 1,
+        FileInformation = 2,
+        Theme = 3
+    }
+
+    #endregion
 }