+++ Source/Krypton Components/Krypton.Toolkit/General/KryptonGlobalToolkitStrings.cs
@@ -10,12 +10,8 @@
 namespace Krypton.Toolkit
 {
     /// <summary>Access 'Global' Krypton string settings.</summary>
-    [Category(@"Code")]
-    [Description(@"Access 'Global' Krypton string settings.")]
-    //[Designer(typeof(KryptonLanguageManagerDesigner))]
-    [ToolboxBitmap(typeof(KryptonLanguageManager), "ToolboxBitmaps.KryptonLanguageManager.bmp")]
-    [ToolboxItem(true)]
-    public class KryptonLanguageManager : Component
+    [TypeConverter(typeof(ExpandableObjectConverter))]
+    public class KryptonGlobalToolkitStrings : GlobalId
     {
         #region Public
 
@@ -33,7 +29,7 @@ namespace Krypton.Toolkit
         /// <summary>Resets the color strings.</summary>
         public void ResetColorStrings() => ColorStrings.Reset();
 
-        /*/// <summary>Gets the button spec style strings.</summary>
+        /// <summary>Gets the button spec style strings.</summary>
         /// <value>The button spec style strings.</value>
         [Category(@"Visuals")]
         [Description(@"Collection of button spec style strings.")]
@@ -45,7 +41,7 @@ namespace Krypton.Toolkit
         private bool ShouldSerializeButtonSpecStyleStrings() => !ButtonStyles.IsDefault;
 
         /// <summary>Resets the button spec style strings.</summary>
-        public void ResetButtonSpecStyleStrings() => ButtonStyles.Reset();*/
+        public void ResetButtonSpecStyleStrings() => ButtonStyles.Reset();
 
         /// <summary>Gets the custom toolkit strings.</summary>
         /// <value>The custom toolkit strings.</value>
@@ -75,76 +71,6 @@ namespace Krypton.Toolkit
         /// <summary>Resets the general strings.</summary>
         public void ResetGeneralStrings() => GeneralToolkitStrings.Reset();
 
-        /*/// <summary>Gets the data grid view style strings.</summary>
-        /// <value>The data grid view style strings.</value>
-        [Category(@"Visuals")]
-        [Description(@"Collection of datagrid view style strings.")]
-        [MergableProperty(false)]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        [Localizable(true)]
-        public DataGridViewStyleStrings DataGridViewStyleStrings => DataGridViewStyles;
-
-        private bool ShouldSerializeDataGridViewStyleStrings() => !DataGridViewStyles.IsDefault;
-
-        /// <summary>Resets the data grid view style strings.</summary>
-        public void ResetDataGridViewStyleStrings() => DataGridViewStyles.Reset();
-
-        /// <summary>Gets the grid style strings.</summary>
-        /// <value>The grid style strings.</value>
-        [Category(@"Visuals")]
-        [Description(@"Collection of grid style strings.")]
-        [MergableProperty(false)]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        [Localizable(true)]
-        public GridStyleStrings GridStyleStrings => GridStyles;
-
-        private bool ShouldSerializeGridStyleStrings() => !GridStyles.IsDefault;
-
-        /// <summary>Resets the grid style strings.</summary>
-        public void ResetGridStyleStrings() => GridStyles.Reset();
-
-        /// <summary>Gets the header group collapsed target strings.</summary>
-        /// <value>The header group collapsed target strings.</value>
-        [Category(@"Visuals")]
-        [Description(@"Collection of header group collapsed target strings.")]
-        [MergableProperty(false)]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        [Localizable(true)]
-        public HeaderGroupCollapsedTargetStrings HeaderGroupCollapsedTargetStrings => GroupCollapsedTargetStrings;
-
-        private bool ShouldSerializeHeaderGroupCollapsedTargetStrings() => !GroupCollapsedTargetStrings.IsDefault;
-
-        /// <summary>Resets the header group collapsed target strings.</summary>
-        public void ResetHeaderGroupCollapsedTargetStrings() => GroupCollapsedTargetStrings.Reset();
-
-        /// <summary>Gets the header style strings.</summary>
-        /// <value>The header style strings.</value>
-        [Category(@"Visuals")]
-        [Description(@"Collection of header style strings.")]
-        [MergableProperty(false)]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        [Localizable(true)]
-        public HeaderStyleStrings HeaderStyleStrings => HeaderStyles;
-
-        private bool ShouldSerializeHeaderStyleStrings() => !HeaderStyles.IsDefault;
-
-        /// <summary>Resets the header style strings.</summary>
-        public void ResetHeaderStyleStrings() => HeaderStyles.Reset();
-
-        /// <summary>Gets the input control style strings.</summary>
-        /// <value>The input control style strings.</value>
-        [Category(@"Visuals")]
-        [Description(@"Collection of input control style strings.")]
-        [MergableProperty(false)]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
-        [Localizable(true)]
-        public InputControlStyleStrings InputControlStyleStrings => InputControlStyles;
-
-        private bool ShouldSerializeInputControlStyleStrings() => !InputControlStyles.IsDefault;
-
-        /// <summary>Resets the input control style strings.</summary>
-        public void ResetInputControlStyleStrings() => InputControlStyles.Reset();*/
-
         /// <summary>Gets the integrated toolbar button strings.</summary>
         /// <value>The integrated toolbar button strings.</value>
         [Category(@"Visuals")]
@@ -156,7 +82,7 @@ namespace Krypton.Toolkit
 
         private bool ShouldSerializeIntegratedToolBarStrings() => !ToolBarStrings.IsDefault;
 
-        /*/// <summary>Resets the integrated tool bar strings.</summary>
+        /// <summary>Resets the integrated tool bar strings.</summary>
         public void ResetIntegratedToolBarStrings() => ToolBarStrings.Reset();
 
         /// <summary>Gets the link behavior style strings.</summary>
@@ -296,7 +222,7 @@ namespace Krypton.Toolkit
         private bool ShouldSerializePaletteImageStyleStrings() => !ImageStyleStrings.IsDefault;
 
         /// <summary>Resets the palette image style strings.</summary>
-        public void ResetPaletteImageStyleStrings() => ImageStyleStrings.Reset();*/
+        public void ResetPaletteImageStyleStrings() => ImageStyleStrings.Reset();
 
         /// <summary>Gets the palette mode strings.</summary>
         /// <value>The palette mode strings.</value>
@@ -312,7 +238,7 @@ namespace Krypton.Toolkit
         /// <summary>Resets the palette mode strings.</summary>
         public void ResetPaletteModeStrings() => ModeStrings.Reset();
 
-        /*/// <summary>Gets the palette text trim strings.</summary>
+        /// <summary>Gets the palette text trim strings.</summary>
         /// <value>The palette text trim strings.</value>
         [Category(@"Visuals")]
         [Description(@"Collection of palette text trim strings.")]
@@ -380,7 +306,35 @@ namespace Krypton.Toolkit
         private bool ShouldSerializeTabStyleStrings() => !TabStyles.IsDefault;
 
         /// <summary>Resets the tab style strings.</summary>
-        public void ResetTabStyleStrings() => TabStyles.Reset();*/
+        public void ResetTabStyleStrings() => TabStyles.Reset();
+
+        /// <summary>Gets the krypton about box basic application information strings.</summary>
+        /// <value>The krypton about box basic application information strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of about box basic application information strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public KryptonAboutBoxBasicApplicationInformationStrings KryptonAboutBoxBasicApplicationInformationStrings => BasicApplicationInformationStrings;
+
+        private bool ShouldSerializeKryptonAboutBoxBasicApplicationInformationStrings() => !BasicApplicationInformationStrings.IsDefault;
+
+        /// <summary>Resets the krypton about box basic application information strings.</summary>
+        public void ResetKryptonKryptonAboutBoxBasicApplicationInformationStrings() => BasicApplicationInformationStrings.Reset();
+
+        /// <summary>Gets the krypton about box strings.</summary>
+        /// <value>The krypton about box strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of about box strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public KryptonAboutBoxStrings KryptonAboutBoxStrings => AboutBoxStrings;
+
+        private bool ShouldSerializeKryptonAboutBoxStrings() => !AboutBoxStrings.IsDefault;
+
+        /// <summary>Resets the krypton about box strings.</summary>
+        public void ResetKryptonAboutBoxStrings() => AboutBoxStrings.Reset();
 
         /// <summary>Gets the krypton miscellaneous theme strings.</summary>
         /// <value>The krypton miscellaneous theme strings.</value>
@@ -410,6 +364,76 @@ namespace Krypton.Toolkit
         /// <summary>Resets the krypton scroll bar strings.</summary>
         public void ResetKryptonScrollBarStrings() => ScrollBarStrings.Reset();
 
+        /// <summary>Gets the data grid view strings.</summary>
+        /// <value>The data grid view strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of data grid view strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public DataGridViewStyleStrings GridViewStyleStrings => DataGridViewStyles;
+
+        private bool ShouldSerializeGridViewStyleStrings() => !DataGridViewStyles.IsDefault;
+
+        /// <summary>Resets the data grid view strings.</summary>
+        public void ResetGridViewStyleStrings() => DataGridViewStyles.Reset();
+
+        /// <summary>Gets the grid style strings.</summary>
+        /// <value>The grid style strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of grid style strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public GridStyleStrings GridStyleStrings => GridStyles;
+
+        private bool ShouldSerializeGridStyleStrings() => !GridStyles.IsDefault;
+
+        /// <summary>Resets the grid style strings.</summary>
+        public void ResetGridStyleStrings() => GridStyles.Reset();
+
+        /// <summary>Gets the header group collapsed target strings.</summary>
+        /// <value>The header group collapsed target strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of header group collapsed target strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public HeaderGroupCollapsedTargetStrings HeaderGroupCollapsedTargetStrings => GroupCollapsedTargetStrings;
+
+        private bool ShouldSerializeHeaderGroupCollapsedTargetStrings() => !GroupCollapsedTargetStrings.IsDefault;
+
+        /// <summary>Resets the header group collapsed target strings.</summary>
+        public void ResetHeaderGroupCollapsedTargetStrings() => GroupCollapsedTargetStrings.Reset();
+
+        /// <summary>Gets the header style strings.</summary>
+        /// <value>The header style strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of header style strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public HeaderStyleStrings HeaderStyleStrings => HeaderStyles;
+
+        private bool ShouldSerializeHeaderStyleStrings() => !HeaderStyles.IsDefault;
+
+        /// <summary>Resets the header style strings.</summary>
+        public void ResetHeaderStyleStrings() => HeaderStyles.Reset();
+
+        /// <summary>Gets the input control style strings.</summary>
+        /// <value>The input control style strings.</value>
+        [Category(@"Visuals")]
+        [Description(@"Collection of input control style strings.")]
+        [MergableProperty(false)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+        [Localizable(true)]
+        public InputControlStyleStrings InputControlStyleStrings => InputControlStyles;
+
+        private bool ShouldSerializeInputControlStyleStrings() => !InputControlStyles.IsDefault;
+
+        /// <summary>Resets the input control style strings.</summary>
+        public void ResetInputControlStyleStrings() => InputControlStyles.Reset();
+
         #endregion
 
         #region Static Strings
@@ -422,6 +446,8 @@ namespace Krypton.Toolkit
         /// <value>The spec style strings.</value>
         public static ButtonStyleStrings ButtonStyles { get; } = new ButtonStyleStrings();
 
+        /// <summary>Gets the custom toolkit strings.</summary>
+        /// <value>The custom toolkit strings.</value>
         public static CustomToolkitStrings CustomToolkitStrings { get; } = new CustomToolkitStrings();
 
         /// <summary>Gets the strings.</summary>
@@ -440,7 +466,7 @@ namespace Krypton.Toolkit
         /// <summary>Gets the group collapsed target strings.</summary>
         /// <value>The group collapsed target strings.</value>
         public static HeaderGroupCollapsedTargetStrings GroupCollapsedTargetStrings { get; } =
-            new HeaderGroupCollapsedTargetStrings();
+        new HeaderGroupCollapsedTargetStrings();
 
         /// <summary>Gets the header styles.</summary>
         /// <value>The header styles.</value>
@@ -473,7 +499,7 @@ namespace Krypton.Toolkit
         /// <summary>Gets the button orientation strings.</summary>
         /// <value>The button orientation strings.</value>
         public static PaletteButtonOrientationStrings ButtonOrientationStrings { get; } =
-            new PaletteButtonOrientationStrings();
+        new PaletteButtonOrientationStrings();
 
         /// <summary>Gets the button spec styles.</summary>
         /// <value>The button spec styles.</value>
@@ -519,10 +545,18 @@ namespace Krypton.Toolkit
         /// <value>The tab styles.</value>
         public static TabStyleStrings TabStyles { get; } = new TabStyleStrings();
 
+        /// <summary>Gets the basic application information strings.</summary>
+        /// <value>The basic application information strings.</value>
+        public static KryptonAboutBoxBasicApplicationInformationStrings BasicApplicationInformationStrings { get; } = new KryptonAboutBoxBasicApplicationInformationStrings();
+
+        /// <summary>Gets the about box strings.</summary>
+        /// <value>The about box strings.</value>
+        public KryptonAboutBoxStrings AboutBoxStrings { get; } = new KryptonAboutBoxStrings();
+
         /// <summary>Gets the miscellaneous theme strings.</summary>
         /// <value>The miscellaneous theme strings.</value>
         public static KryptonMiscellaneousThemeStrings MiscellaneousThemeStrings { get; } =
-            new KryptonMiscellaneousThemeStrings();
+        new KryptonMiscellaneousThemeStrings();
 
         /// <summary>Gets the scroll bar strings.</summary>
         /// <value>The scroll bar strings.</value>
@@ -532,8 +566,8 @@ namespace Krypton.Toolkit
 
         #region Identity
 
-        /// <summary>Initializes a new instance of the <see cref="KryptonLanguageManager" /> class.</summary>
-        public KryptonLanguageManager()
+        /// <summary>Initializes a new instance of the <see cref="KryptonGlobalToolkitStrings" /> class.</summary>
+        public KryptonGlobalToolkitStrings()
         {
             //throw new NotImplementedException();
         }
@@ -543,44 +577,45 @@ namespace Krypton.Toolkit
         #region Implementation
 
         /// <summary>Gets a value indicating whether this instance is default.</summary>
-        /// <value>
-        ///   <c>true</c> if this instance is default; otherwise, <c>false</c>.</value>
+        /// <value><c>true</c> if this instance is default; otherwise, <c>false</c>.</value>
         [EditorBrowsable(EditorBrowsableState.Never)]
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
         public bool IsDefault => !(ShouldSerializeCustomStrings() ||
-                                   ShouldSerializeGeneralStrings() ||
-                                   ShouldSerializeColorStrings() ||
-                                   ShouldSerializePaletteModeStrings() ||
-                                   //ShouldSerializeButtonSpecStyleStrings() ||
-                                   //ShouldSerializeDataGridViewStyleStrings() ||
-                                   //ShouldSerializeGridStyleStrings() ||
-                                   //ShouldSerializeHeaderGroupCollapsedTargetStrings() ||
-                                   //ShouldSerializeHeaderStyleStrings() ||
-                                   //ShouldSerializeInputControlStyleStrings() ||
-                                   ShouldSerializeIntegratedToolBarStrings() ||
-                                   //ShouldSerializeKryptonLinkBehaviorStrings() ||
-                                   //ShouldSerializePaletteBackStyleStrings() ||
-                                   //ShouldSerializePaletteBorderStyleStrings() ||
-                                   //ShouldSerializePaletteButtonOrientationStrings() ||
-                                   //ShouldSerializePaletteButtonSpecStyleStrings() ||
-                                   //ShouldSerializePaletteButtonStyleStrings() ||
-                                   //ShouldSerializePaletteContentStyleStrings() ||
-                                   //ShouldSerializePaletteImageEffectStrings() ||
-                                   //ShouldSerializePaletteImageStyleStrings() ||
-                                   //ShouldSerializePaletteTextTrimStrings() ||
-                                   //ShouldSerializePlacementModeStrings() ||
-                                   //ShouldSerializeSeparatorStyleStrings() ||
-                                   //ShouldSerializeTabBorderStyleStrings() ||
-                                   //ShouldSerializeTabStyleStrings() ||
-                                   ShouldSerializeKryptonMiscellaneousThemeStrings() ||
-                                   ShouldSerializeKryptonScrollBarStrings());
+       ShouldSerializeGeneralStrings() ||
+       ShouldSerializeColorStrings() ||
+       ShouldSerializePaletteModeStrings() ||
+       ShouldSerializeButtonSpecStyleStrings() ||
+       ShouldSerializeGridViewStyleStrings() ||
+       ShouldSerializeGridStyleStrings() ||
+       ShouldSerializeHeaderGroupCollapsedTargetStrings() ||
+       ShouldSerializeHeaderStyleStrings() ||
+       ShouldSerializeInputControlStyleStrings() ||
+       ShouldSerializeIntegratedToolBarStrings() ||
+       ShouldSerializeKryptonLinkBehaviorStrings() ||
+       ShouldSerializePaletteBackStyleStrings() ||
+       ShouldSerializePaletteBorderStyleStrings() ||
+       //ShouldSerializePaletteButtonOrientationStrings() ||
+       ShouldSerializePaletteButtonSpecStyleStrings() ||
+       ShouldSerializePaletteButtonStyleStrings() ||
+       ShouldSerializePaletteContentStyleStrings() ||
+       ShouldSerializePaletteImageEffectStrings() ||
+       ShouldSerializePaletteImageStyleStrings() ||
+       ShouldSerializePaletteTextTrimStrings() ||
+       ShouldSerializePlacementModeStrings() ||
+       ShouldSerializeSeparatorStyleStrings() ||
+       ShouldSerializeTabBorderStyleStrings() ||
+       ShouldSerializeTabStyleStrings() ||
+       ShouldSerializeKryptonAboutBoxBasicApplicationInformationStrings() ||
+       ShouldSerializeKryptonAboutBoxStrings() ||
+       ShouldSerializeKryptonMiscellaneousThemeStrings() ||
+       ShouldSerializeKryptonScrollBarStrings());
 
         /// <summary>Resets this instance.</summary>
         public void Reset()
         {
             ResetColorStrings();
 
-            //ResetButtonSpecStyleStrings();
+            ResetButtonSpecStyleStrings();
 
             ResetCustomStrings();
 
@@ -588,45 +623,49 @@ namespace Krypton.Toolkit
 
             ResetPaletteModeStrings();
 
-            //ResetDataGridViewStyleStrings();
+            ResetGridViewStyleStrings();
 
-            //ResetGridStyleStrings();
+            ResetGridStyleStrings();
 
-            //ResetHeaderGroupCollapsedTargetStrings();
+            ResetHeaderGroupCollapsedTargetStrings();
 
-            //ResetHeaderStyleStrings();
+            ResetHeaderStyleStrings();
 
-            //ResetInputControlStyleStrings();
+            ResetInputControlStyleStrings();
 
-            //ResetIntegratedToolBarStrings();
+            ResetIntegratedToolBarStrings();
 
-            //ResetKryptonLinkBehaviorStrings();
+            ResetKryptonLinkBehaviorStrings();
 
-            //ResetPaletteBackStyleStrings();
+            ResetPaletteBackStyleStrings();
 
-            //ResetPaletteBorderStyleStrings();
+            ResetPaletteBorderStyleStrings();
 
             //ResetPaletteButtonOrientationStrings();
 
-            //ResetPaletteButtonSpecStyleStrings();
+            ResetPaletteButtonSpecStyleStrings();
 
-            //ResetPaletteButtonStyleStrings();
+            ResetPaletteButtonStyleStrings();
 
-            //ResetPaletteContentStyleStrings();
+            ResetPaletteContentStyleStrings();
 
-            //ResetPaletteImageEffectStrings();
+            ResetPaletteImageEffectStrings();
 
-            //ResetPaletteImageStyleStrings();
+            ResetPaletteImageStyleStrings();
 
-            //ResetPaletteTextTrimStrings();
+            ResetPaletteTextTrimStrings();
 
-            //ResetPlacementModeStrings();
+            ResetPlacementModeStrings();
 
-            //ResetSeparatorStyleStrings();
+            ResetSeparatorStyleStrings();
 
-            //ResetTabBorderStyleStrings();
+            ResetTabBorderStyleStrings();
 
-            //ResetTabStyleStrings();
+            ResetTabStyleStrings();
+
+            ResetKryptonKryptonAboutBoxBasicApplicationInformationStrings();
+
+            ResetKryptonAboutBoxStrings();
 
             ResetKryptonMiscellaneousThemeStrings();